[TechDraw] Move CenterLine into own file
