+++ src/Mod/TechDraw/App/CMakeLists.txt
@@ -184,6 +184,8 @@ SET(Geometry_SRCS
     Geometry.h
     GeometryObject.cpp
     GeometryObject.h
+    CenterLine.cpp
+    CenterLine.h
     Cosmetic.cpp
     Cosmetic.h
     CosmeticVertex.cpp
