+++ src/Mod/TechDraw/App/CosmeticExtension.cpp
@@ -24,6 +24,7 @@
 
 #include <Base/Console.h>
 
+#include "CenterLine.h"
 #include "CosmeticExtension.h"
 #include "CosmeticExtensionPy.h"
 #include "Cosmetic.h"
